backend/requirements.txt
pyswisseph==2.10.3.2
fastapi==0.115.2
uvicorn[standard]==0.30.6
pytz==2024.1
numpy==2.1.1
python-dateutil==2.9.0.post0